{{ cookiecutter.project_name }}/requirements_dev.txt
# Dev
pip
bumpversion
pytest
build
twine
pyyaml
pymongo
runscript
coverage
pytest-cov
pytest-xdist
rich

# types
types-pyyaml

# types
types-pyyaml

# Code Quality
ruff
flake8
mypy
pep8-naming
pylint
msgspec
black
